+++ src/web/BitmapFont.ts
@@ -1,4 +1,5 @@
 import { FontJSON } from "*.font.json";
+import { DisplayManager, RenderMode } from "./DisplayManager";
 import { getRenderingContext, loadImage } from "./graphics";
 
 const CHAR_SPACING = 1;
@@ -121,14 +122,16 @@ export class BitmapFont {
         ctx.save();
         ctx.translate(x, y);
 
-        // Ugly hack to correct text position to exact pixel boundary because Chrome renders broken character images
-        // when exactly between two pixels (Firefox doesn't have this problem).
-        if (ctx.getTransform != null) {
-            const transform = ctx.getTransform();
-            ctx.translate(
-                Math.round(transform.e) - transform.e,
-                Math.round(transform.f) - transform.f
-            );
+        if (DisplayManager.getInstance().getRenderMode() !== RenderMode.NATIVE) {
+            // Ugly hack to correct text position to exact pixel boundary because Chrome renders broken character images
+            // when exactly between two pixels (Firefox doesn't have this problem).
+            if (ctx.getTransform != null) {
+                const transform = ctx.getTransform();
+                ctx.translate(
+                    Math.round(transform.e) - transform.e,
+                    Math.round(transform.f) - transform.f
+                );
+            }
         }
 
         ctx.globalAlpha *= alpha;