+++ src/web/scenes/OptionsScene.ts
@@ -9,14 +9,17 @@ import { Direction } from "../geom/Direction";
 import { ControllerEvent } from "../input/ControllerEvent";
 import { TextNode } from "../scene/TextNode";
 import { BitmapFont } from "../BitmapFont";
-import { MenuItem, MenuList, SliderMenuItem } from "../Menu";
+import { MenuItem, MenuList, SelectMenuItem, SliderMenuItem } from "../Menu";
 import { ControlTooltipNode } from "../scene/ControlTooltipNode";
 import { ControllerAnimationTags } from "../input/ControllerFamily";
 import { AudioManager } from "../audio/AudioManager";
 import { SoundChannel } from "../audio/SoundChannel";
+import { DisplayManager, RenderMode } from "../DisplayManager";
+import { isDev } from "../util";
 
 enum MenuItemKey {
     FULLSCREEN = "fullscreen",
+    RENDER_MODE = "render-mode",
     SFX_SLIDER = "sfxSlider",
     MUSIC_SLIDER = "musicSlider",
 }
@@ -86,6 +89,26 @@ export class OptionsScene extends Scene<FriendlyFire> {
             new MenuItem(
                 MenuItemKey.FULLSCREEN, "Toggle Fullscreen", OptionsScene.font, "black", menuItemX, menuItemY
             ),
+            new SelectMenuItem(
+                {
+                    id: MenuItemKey.RENDER_MODE,
+                    label: "Render Mode",
+                    font: OptionsScene.font,
+                    color: "black",
+                    x: menuItemX,
+                    y: menuItemY + 20,
+                    enabled: true,
+                    initialValue: this.game.displayManager.getRenderMode(),
+                    values: Object.values(RenderMode).filter(v => v !== RenderMode.NATIVE || isDev()),
+                    valueLabels: {
+                        [ RenderMode.PIXEL_IMPERFECT ]: "Pixel Imperfect",
+                        [ RenderMode.PIXEL_PERFECT ]: "Pixel Perfect",
+                        [ RenderMode.NATIVE ]: "Native (Experimental)",
+                    },
+                    actionCallback: this.handleRenderModeChange,
+                    data: { displayManager: this.game.displayManager }
+                }
+            ),
             new SliderMenuItem(
                 {
                     id: MenuItemKey.SFX_SLIDER,
@@ -93,7 +116,7 @@ export class OptionsScene extends Scene<FriendlyFire> {
                     font: OptionsScene.font,
                     color: "black",
                     x: menuItemX,
-                    y: menuItemY + 20,
+                    y: menuItemY + 40,
                     enabled: true,
                     initialValue: Math.round(this.audioManager.sfxGain * 100),
                     minValue: 0,
@@ -111,7 +134,7 @@ export class OptionsScene extends Scene<FriendlyFire> {
                     font: OptionsScene.font,
                     color: "black",
                     x: menuItemX,
-                    y: menuItemY + 40,
+                    y: menuItemY + 60,
                     enabled: true,
                     initialValue: Math.round(this.audioManager.musicGain * 100),
                     minValue: 0,
@@ -124,6 +147,10 @@ export class OptionsScene extends Scene<FriendlyFire> {
             )
         );
         this.menu.appendTo(panel);
+    }
+
+    private handleRenderModeChange(newValue: RenderMode, data: { displayManager: DisplayManager }): void {
+        data.displayManager.setRenderMode(newValue);
     }
 
     private handleAudioSliderChange(newValue: number, data: { channel: SoundChannel, audioManager: AudioManager }): void {